feat(bsc): add -print-bs-libdir flag
Previous versions of the Bluespec compiler required you to specify the
`BLUESPECDIR` environment variable to locate the compiler toolchain.
That's no longer necessary, which is great, but there is another use
case provided by that variable: it makes it easy to find the necessary
Verilog files to add to a Bluespec project. Tools such as `yosys-bsv`
rely on functionality like this, so they can automatically synthesize
working designs out of the box.

While there's various ways to coax this out of the compiler, this new
`-print-bs-libdir` flag just does one thing only: prints out the path to
the `lib/` subdirectory, and nothing else. The intention is essentially
to use it for scripting purposes, to find the proper library directory
of `bsc` without a lot of shennanigans. Putting this functionality in
`bsc` also makes it more robust: if the semantics of install dirs ever
change (e.g. a concrete `--prefix` flag), scripts using this feature
will keep working regardless.

Signed-off-by: Austin Seipp <aseipp@pobox.com>