don't build targets in `comp` in parallel
Otherwise, src/comp/Makefile invokes multiple ghc builds, all in the
same workspace, which break each other sometimes.

One way of still getting some parallelization there would be to pass -jN
to GHC itself.

Reading what number of jobs is passed to make, and passing it down to
GHC doesn't really work
[without a lot of hackery](https://stackoverflow.com/questions/5303553),
and there seems to be a
[upper limit](https://gitlab.haskell.org/ghc/ghc/issues/9221) as well,
so I decided to introduce a GHCJOBS=n variable.

However, compilation seems to be somewhat unstable, which is why it's
defaulted it to 1 for now (and not a "reasonable" number like 4).

Fixes #38