Fix BSC compile when using Clang.
Update gcc-script.sh to support clang.

Use script as GHC assembler, and linker as well as just the compiler.
This is necessary for shared libraries to link properly
with ghc when invoed with make CC=clang-6.0 CXX=clang++-6.0

Note: Explicitly do *not* use clang as the preprocessor yet, as it's
stricter than GCC's and fails on some macros in the compiler source.
We'll need to fix this for MacOS X support.