Add a check of the dynamically linked STP library
Part of checking the command-line flags is checking that the chosen
SAT solver is available.  A check was only being done when Yices was
selected, since the real Yices library wasn't being shipped with BSC.
When a valid Yices was not found, it was reported as a warning and
the flags were updated to choose STP as the solver (since that was
assumed to always be available).  In preparation for allowing STP to
also be a stub, add a check of the STP library when it's selected,
and make the warning be an error (regardless of the chosen solver),
and improve the error message.  For STP, since there's no functions
in the current API for querying the version, check for a valid
library by testing whether the create function returns a non-null
pointer.