ci(github): touchup CI workflows
This new workflow splits out the whitespace check into a separate job in
the workflow. This helps improve parallelism and accuracy (since we
don't run the check once for every OS in the build matrix), and also
puts "code checks" into its own category in the Workflow output.

It also gives all of the CI jobs consistent names, which are `Build:
<FOO>` where `<FOO>` is generally the name of the system/task that's
being run.

Signed-off-by: Austin Seipp <aseipp@pobox.com>