Compile on MacOS
Uses builtin tcl/tk/itcl/itk but gcc, ghc, icarus from Homebrew.
Homebrew GCC is only used for STP; everything else is compiled
with the Xcode compiler.

In make check, bluetcl crashes at the end of simulation with
"Called Tcl_FindHashEntry on deleted table". This seems to be
an issue with the system Tcl.