Explicitly check itk ver instead of assuming it's the same as itcl
NB this will require xvfb to be installed on Linux both for setups
with no X server running, and for ones with an X server so we avoid the
build briefly displaying a wish window and stealing focus.

On macOS we explicitly use the system version of wish and tclsh in
/usr/bin and the system tcl/tk header files.

Update README and github build scripts to take this into account.