Remove compiler-script.sh command line filter.
Realistically, we don't care about the versions of GCC it supports
(GCC <4.4), because we require C++11 support (GCC >=4.8) elsewhere,
and it causes some versions of GHC on Linux to break on certain
systems.

In particular it interacts badly with systems where -fPIE is the
C toolchain default and the GHC setting "C compiler supports-no-pie"
is "YES", because it stops GHC passing -no-pie as CFLAGS and LDFLAGS.

See:
https://gitlab.haskell.org/ghc/ghc/issues/17919
https://github.com/B-Lang-org/bsc/issues/109

Thanks to @leonschoorl for tracking this down.