Load SAT libs from dir relative to exe using linker flags
This allows us to remove the LD_LIBRARY_PATH code from the
wrapper scripts (step 1 of removing the scripts altogether).

On macOS, this requires us to prefix @rpath/ to libraries'
install_names.

Then on macOS and Linux respectively, we can link with
-rpath @executable_path/../../lib/SAT and $ORIGIN/../../lib/SAT
respectively.

While here, fix up the file naming for the stub libyices on macOS:
It was using the Linux shared library names.