Silence false positive overflow error on gcc 10
`0 <= numDigits(v, base) <= 64` for any `v`, `base`. However, since the
value is returned as an `unsigned int`, gcc 10 alerts for a possible
overflow on `char buf[value_width+1]`, although it is a false positive.
By returning an `unsigned short`, the warning (which is actually treated
as an error due to the CXXFLAGS) is silenced.