Update BSC's support for ModelSim/Questa
Add support for "-vsim questa", using the same exec script as ModelSim,
since they share the same front end interface.

Remove the "-novopt" flag, which was deprecated and then removed.
It was being used to preserve access to all signals, for VCD dumping.
The vopt stage is now always performed and the way to preserve
visibility is with the "+acc" flag (which can take an argument, or
preserves access to the entire design when used without an argument).