Add a -(no-)show-version arg to bsc.
This removes the version from the "Generated by Bluespec Compiler"
comment added to generated verilog and C++.

This is useful for build systems (such as bazel, and buck) where
rebuilds are determined by the hash of the tool inputs. If we can
avoid rebuilds of the compiler itself from unnecessarily causing
source code changes, it will improve the effectiveness of the build
cache.

This will be also be useful for the bluespec test suite which builds
a lot of verilog and bluesim (C++) simulations, and it will help if
we can reuse the same C++ build cache across multiple builds of the
compiler itself.