Silence false positive overflow error on gcc 10
`0 <= numDigits(v, base) <= 64` for any `v`, `base`. However, since the
value is returned as an `unsigned int`, gcc 10 alerts for a possible
overflow on `value_width+1`, because of C++'s integer promotion rules
which won't promote a (usually 32-bit) int to a (usually 64-bit) size_t
on LP64 platforms where pointers and array indices are 64 bit.