Fix the type of the 'guarded' param on FIFO prims
This parameter indicates whether a FIFO's methods should be guarded or
unguarded and therefore is Bool, not Integer.  Further, parameters of
Integer type become 32-bit values in the generated Verilog's module
instantiation; but the Verilog primitives are using the parameter as
an argument to boolean logic operators, so there is a size mismatch
(flagged by the Verilator linter).  This commit fixes the instantiations
to provide a 1-bit value for the parameter.