Support TCLLIBPATH env variable in bluetcl
The standard Tcl init will handle TCLLIBPATH, but only if the Bluetcl
app doesn't set auto_path first; so adjust the auto_path later.  And
instead of doing that in the app code, do that in the tcl script that
the app code is already sourcing.  And improve the comments.