Reinstate the 'exit_status' test proc
This exposes the internal 'exec_with_log_and_return' as a test proc
that runs and command and determines pass/fail based on the exit
status.  This is not currently used in the testsuite, but may be
useful for the future, or can be used by other repos that want to
define their own tests that use this infrastructure.