prims: consolidate vivado/generic Verilog primitives
The Verilog primitives for Vivado are identical to the generic
primitives for all synthesis tools, except they have some Verilog
annotations for the clocking and RAM primitives to use the right
underlying logic. So all of these can be consolidated by just using
`ifdef appropriately. Do this, and remove all the copies.

Signed-off-by: Austin Seipp <aseipp@pobox.com>