Testsuite: Combine parser directories
Merge the two BH/Classic directories into one (bsc.cparse and bsc.syntax);
and move it, along with the two BSV directories, under a single hierarchy.
Also, rename 'bsv05-parse-pretty' to use underscores instead of hyphens,
because it was causing 'make check' run in 'bsc.syntax' to re-run the
subdirectories of 'bsv05', for some reason.