Let CStruct encode struct vs constructor when known
In BH/Classic, the same syntax is used for expressing structs and
constructors with named fields.  The BH/Classic parser represents both
with CStruct and leaves it to the typechecker to disambiguate.  In BSV,
the syntax indicates which is intended, but that information was being
lost.  This has been fixed by adding a field to CStruct to allow the
BSV parser to record whether a CStruct is known to be a struct or a
constructor.  There are also CStruct expressions created inside BSC
where the meaning is known, and that can now also be recorded.
This reduces the work that the typechecker needs to do and avoids
mistakenly interpreting a CStruct in the wrong way.

This change uncovered that a number of BSV files in the testsuite were
using the wrong syntax.  These have been fixed by adding or removing
the keyword 'tagged' as needed.  New tests have been added, to test the
current state of the parsers and typechecker, testing this change and
other issues raised in bug #353.