Further improve typecheck of CStruct
The previous commit removed the need to disambiguate CStruct value
expressions created by the BSV parser.  CStruct values created by the
BH/Classic parser still need to be disambiguated.  This commit
improves the heuristic for that.  The BH/Classic test for bug #353 now
compiles.

This commit also disallows using CStruct for constructors without
named fields.  A new error message is given for this situation.  Tests
for this situation are updated.  The tests reveal that non-named
fields can still be mentioned in pattern matching, which will need to
be fixed.

It turns out that typechecking of fields in patterns (CPstruct) is
done by different code in BSC, so the improvements so far have not
addressed patterns.  New test examples are added, that show that the
original disambiguation issue still exists for CPstruct in pattern
matches.