Improve error msgs for parsing of tagged field patterns
by removing a use of the 'try' function.  Update the testcase.

Disallow parsing of empty tuple braces.  (We might even want to
disallow braces with only one pattern, which is currently interpretted
as ignoring the braces.)

Add more testing of tuples in patterns -- as top-level patterns and
as the pattern of tagged unions.

Add more testing of anonymous struct patterns (in tagged unions), to
include testing when not all fields are mentioned, including when no
fields are given (just empty braces).