Expand the CI's trailing whitespace check
to also check for trailing tabs and to also find trailing space or tab
in DOS files where the space/tab is not technically trailing, because
there is a carriage return (\r) at the end of the line -- which is
trivially handled by disallowing DOS files and checking for any
trailing space, including CRs and line feeds (LF).  This uncovered a
number of files that had trailing tabs, LFs, or were in DOS format
(and in some cases had trailing spaces), that have been updated.
Allow some files to be exempted with a file of patterns to filter.