Clarify the download and install processes
Only mention downloading (of pre-built releases) on the entry README,
and move the "build from source" documentation to its own file, with
additional clarification about the result of the build (the 'inst'
directory) and how to install it.  Add some of this clarification to
the Makefile as well.  And install a README into the 'inst' release,
clarifying how to use the directory.