Update User Guide flags to match help message
and add perl scripts that can be used to check for differences between
the LaTeX source and BSC's help message (the -help output for BSC and
the -h output for Bluesim executables).  These scripts found that some
newer flags had not yet been documented and that one flag description
had changed.