Remove the CI's claim to build for macOS 10.13+
The build contains many warnings about linking with objects that were
built for a newer version than being targeted.  That's partly because
we ran 'brew' and 'cabal' to install dependencies and we didn't specify
the target there.  But even if we fixed this, there are other homebrew
libraries pre-installed in the VM that we don't have control over --
specifically 'libgmp'.  Maybe we could force reinstall of this library,
but it seems best to just give up on building a 10.13+ build.