Cleanup/update the Makefile code for invoking GHC
Since we no longer support GHC versions before 7.10, the Makefile does
not need to include ifdef directives for earlier versions.  The flags
and packages since 7.10 have stabilized, so a lot of conditional code
is not needed.

Also, report an error if the GHC version is lower than supported.
And, report a warning if the version if newer than expected, since the
source will not yet have been tested for that version.
Finally, add a workaround for a bug in GHC 9.2.1, that turns off
optimization; BSC builds and passes the testsuite, but there are many
new warnings that need investigating.