Update Verilator link script
Fix to handle when Verilog modules (and in particular the top-level
module) are provided by BSC.

Add support for DPI (will need support from BSC).

Detect whether 'main.v' is provided (in prep for supporting no 'main').

Fix a variable name which had 'IV' in it, due to copying from the
iverilog script.