Fix remaining issues with DPI codegen
Verilator does not support open arrays, so polymorphic functions will
not work; but all the other foreign import tests are working.
Wide return values are now passed as an 'output' first argument.
All other arguments are declared as 'input' to avoid misinterpretation
(and allow constant values to be passed).
Fix a typo in 'longint' for values of bit size 33 to 64.

For system tasks without arguments, BSC does not print the empty
parentheses, but system tasks are syntactically distinguishable by
the initial dollar sign in the name.  For imported DPI functions,
without parentheses, it looks like a variable and not a function call,
and Verilator complains.  So have BSC include the empty parentheses
in that case.