Fix DPI support to declare in the Verilog files
The link stage passes '-dpi' to the vsim link script, to indicate to
use DPI, instead of the script looking for a 'dpi_declaration.v' file.
The Modelsim/Questa script is updated to work with DPI; along with the
Verilator script.  All other scripts now detect the flag and error
(that they don't yet support DPI).

The changes to the VProgram type are currently a draft -- the String
field should possibly be an abstract type.

The DPI declarations are updated to declare byte/int/longint as unsigned.