Specify the top module name to Verilator
If no name is specified, Verilator uses the name of the first file
on the command line.  But OVL tests in the testsuite put a ".vlib"
first.  By explicitly giving the name, those tests will compile.
We also specify the top module, in case some other files provided
on the command-line define modules not instantiated in the design.