Testsuite: Remove the fragile use of MAKELEVEL
The file 'all_tests.mk' is only needed for the 'run-tests' target, that
is only need for parallel testing.  But it always needed to be imported,
and code was added to conditionally create it, using MAKELEVEL to attempt
to create it only once.  This fails, though, when entering the testsuite
makefile from another makefile (such as the 'check-suite' target in the
top directory).  The testsuite makefiles should really be cleaned up
(if not replaced with something better); but for now, we can make the
current targets work by putting the code for parallel tests (like the
include of 'all_tests.mk') into a separate 'parallel.mk' that is only used
when making the parallel targets.