Draft support for DPI-based BDPI and Verilator vsim linking (#439)
Add a vsim script for Verilator.

Remove a duplicate cvc64 link script.

Add a new flag -use-dpi, that needs to be given during both compile
and link steps.  This flag is "hidden" and undocumented, which can be
changed when this has been further tested and fleshed out.  When a
Verilog module is generated, if it uses any DPI calls, declarations
for them are included at the top of the file containing the module.

The vsim scripts now look for `-dpi` on their command-line as an
indication to link for DPI.  The ModelSim/Questa and Verilator scripts
support this flag, while the other scripts give an error that it is
not supported.

Polymorphic BDPI is not yet supported with DPI.  BSC will give an
error if a compiled module for Verilog uses a polymoprhic BDPI
function.

Add some support in the testsuite for running with Verilator as the
simulator.  Verilator finishes execution of blocks (or timesteps?)
after $finish is called, so there are some displays that may appear,
that need to be filtered out when comparing against expected.  This
only works for displays; some examples write to files after $finish is
called, and theres currently no filtering of those extra lines.