Fix ifc schedule for sized Bypass FIFOs and add tests
The notEmpty/notFull methods were conflict-free with enq/deq, which
meant that they could be called in any order, but the values are only
valid when called before enq/deq (for notFull) or after enq before deq
(for notEmpty).  The test for Bypass FIFO schedules was exempting the
sized version because it has a dedicated test, but that test was not
catching this issue.