Resolve CI warnings by updating to newer action versions
Update checkout, cache, download-artifact, and upload-artifact
from v2 to v3.  There were some uses of upload-artifact@v1, but I can
see no reason why we held back to the older version (and the choice
wasn't consistent!), so these have been updated to v3.

In both the ubuntu and macos jobs for building the docs, there are
separate actions for uploading a tarfile vs a directory.  Comments
were only included on the ubuntu job, so this adds them to the macos
job, and also attempts to make the wording more clear.