Extend the tests for logic between system tasks
to include an example when the logic is in a separately synthesized
submodule.  For Verilog, system tasks are executed in an always block
and zero-delay statements are used to ensure that assignments have a
chance execute between tasks.  BSC now inlines assignment statements
from the current module, so the delay is no longer needed in those
situations.  This commit adds an example where the delay is still
necessary.  It also adds an example where inlined logic appears before
the outside logic, revealing a bug in where BSC inserts the delay
statement.