Add TEST_SYSTEMC_CXXFLAGS option to testsuite
And use it to get the SystemC tests passing on macOS, now that the
SystemC library installed by Homebrew requires a flag to the C++ compiler.

Also, cleanup how the SystemC flags are received in the testsuite.
The testsuite has "which_*" procedures for getting the location of
executables and "get_*" procedures for getting other values -- the
SYSTEMC_INC and SYSTEMC_LIB variables were incorrectly using "which_*".
The testsuite requires that these be set (if SystemC tests are enabled)
and they can be expected to exist as variables, so there's no need for
"get_*" functions for them; instead the functionality is bundled into
a single function that is called at the start and that installs the
values into global values -- in the same way that "get_test_options"
does for "vtest" etc (and the SystemC options could eventually be
bundled in with that).  The SystemC test procedures can also assume
that the values exist, so conditional code to check for that is removed.