Update iverilog linking with VPI for v11+
In IVerilog v11+, system function types (sft) are taken directly from the VPI files. Previously, a separate .sft file had to be generated.  Newer IVerilog accepts a .sft file, but emits a deprecation warning.  This warning caused unexpected output for one test in the testsuite, that was addressed by filtering out the warning; that temporary fix is now removed.

The -L flag allows linking the VPI files as a library for type checking.

A fix is also applied to a guard on the IVerilog version, on an unrelated test.  The verilog compiler version string may contain more than just major and minor version numbers, so a direct comparison to the string should not be made; instead, we extract the maj.min pattern before comparison.