Generate VPI files in the same dir with Verilog
Generated Verilog files are placed in the -vdir directory or, if
no -vdir is specified, in the same directory as the source file being
compiles.  VPI wrapper files are also placed in the -vdir, but if
no -vdir is specified, they were being placed in the current directory
(where BSC is invoked).  These should be consistent, so the location
for VPI wrappers is fixed to match the generated Verilog.