GitHub CI: upgrade from ghc 9.0.2 to ghc 9.2.8
We held back to GHC 9.0 because of a bug in 9.2.1, but that has been
fixed.  GHC 9.2.8 is the current recommended version by GHCUP.
The CI output will now contain many new warnings, since GHC 9.2
warns about more issues -- this is okay, but we should work on
resolving them (see GitHub issue 469).