Update BSC to compile with GHC 9.6
In version 2.3 of the mtl library, re-export of other modules was
removed; so now those other modules need to be imported, if functions
from them are needed.  To avoid duplicate import when compiling with
earlier mtl/GHC versions, the import statements need to explicitly
list the functions imported from each module.