Introduce STP_USE_LIB flag
The new build flag STP_USE_LIB can be set to choose the STP library to be used.
Options are '', 'included', 'system', 'disable', and an absolute path starting with '/'.

The empty string defaults to the same behavior as 'included' which builds STP from the provided source.
'disable' has the same effect as STP_STUB=1, i.e. disabling STP at runtime.
'system' tries to find the system library (currently only in /usr/lib) and then copies it for building.
An absolute path starting with '/' can be provided and must point to the directory in which the desired STP library is located. Otherwise the
same as 'system'.

The options for precompiled STP libraries essentially do the same things as STP_STUB, except they copies the library
instead of compiling the stub. For compatibility reasons within the Makefiles, the library is linked to under
the soname libstp.so.1 eventhough technically it should be libstp.so.2.3 (for the newest versions), but STP is backwards-compatible
enough that this does not cause issues.

Enabling the flag reduced the build time by up to 25% (the time to compile STP from source), while `make TEST_BSC_OPTIONS="-sat-stp" smoke`
saw no reductions, both compared to the provided source and Yices.